backend/requirements.txt
click==8.1.7
Flask==2.0.1
flask-cas-ng==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
MarkupSafe==2.1.4
Werkzeug==2.0.1
xmltodict==0.13.0
